ProfileView 컴포넌트 로직 auth context로 분리
